Redo multithreading to use more threads more of the time. Previously we would start a group of threads and then wait for them all to finish before starting the next group, but this is very inefficient if e.g. most of the threads take 5 minutes and one takes 10 minutes.
Now we use a `Mutex<bool>` to indicate whether threads should be checking for new work to do, and a condition variable to alert threads that aren't currently doing anything. When one thread `T` finds work to do, `T` notifies one other thread to look for work, and then does so again once `T`s work is finished.

We also hold the current state of each chunk/set of update files in an `Arc<Mutex<Vec<_>>>` so that each thread is guaranteed to only do work on parts of the data that aren't already being touched by another thread.