Change `is_writing` condition variable to `is_writing_all`, and allow `write_from_source` to run at the same time as it, but wait for calls to `write_all` to finish before returning from `update_and_expand_chunk`.
Also, add a "priority" condition variable for threads that just need to wait for `write_all` to finish, but don't want to call `write_all` themselves.

We need to do this because in the current megaminx <U,R> search, it appears that some threads are blocking on the call to `write_from_source` for over 2 hours. Presumably this is because they are waiting for `is_writing`, but other threads keep getting the mutex lock first. After notifying the priorty condvar, we sleep for 1 second before notifying the regular condvar, which should allow enough time for threads blocked on the priority condvar to wake up and proceed.